requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blessed==1.20.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
editor==1.6.6
faiss-cpu==1.10.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inputimeout==1.0.4
inquirer==3.4.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-text-splitters==0.3.6
langsmith==0.3.7
Levenshtein==0.26.1
lib_resume_builder_AIHawk==0.2
loguru==0.7.3
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
# Se ha eliminado la línea que fijaba numpy==2.2.2
openai==1.61.1
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.12.1
readchar==4.2.1
regex==2024.7.24
reportlab==4.3.0
requests==2.32.3
requests-toolbelt==1.0.0
runs==1.2.2
selenium==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xmod==1.8.1
yarl==1.18.3
zstandard==0.23.0

